IOS-6262 Fix review
